Big Release!!! Create client using metro ui , mvvm and mvvm light library. Implement base DB loading using DAL and entity repo. Refactoring. Rewrite mainwindow in Command_Calc. Fix EM bag and some bags in repositories. Add entity servises

+++ Octagon/DAL/EntityModel/UserMemory.cs
@@ -20,7 +20,7 @@ namespace DAL.EntityModel
         }
     
         public int User_ID { get; set; }
-        public Nullable<decimal> Memory { get; set; }
+        public Nullable<double> Memory { get; set; }
     
         public virtual ICollection<UserHistory> UserHistories { get; set; }
     }
